// client/stocklinx app/src/redux/deployedProduct/constant.ts
export enum deployedProductConst {
  FETCH_DEPLOYEDPRODUCTS_REQUEST = "FETCH_DEPLOYEDPRODUCTS_REQUEST",
  FETCH_DEPLOYEDPRODUCTS_SUCCESS = "FETCH_DEPLOYEDPRODUCTS_SUCCESS",
  FETCH_DEPLOYEDPRODUCTS_FAILURE = "FETCH_DEPLOYEDPRODUCTS_FAILURE",
  FETCH_DEPLOYEDPRODUCT_REQUEST = "FETCH_DEPLOYEDPRODUCT_REQUEST",
  FETCH_DEPLOYEDPRODUCT_SUCCESS = "FETCH_DEPLOYEDPRODUCT_SUCCESS",
  FETCH_DEPLOYEDPRODUCT_FAILURE = "FETCH_DEPLOYEDPRODUCT_FAILURE",
  FETCH_DEPLOYEDPRODUCT_COUNTS_REQUEST = "FETCH_DEPLOYEDPRODUCT_COUNTS_REQUEST",
  FETCH_DEPLOYEDPRODUCT_COUNTS_SUCCESS = "FETCH_DEPLOYEDPRODUCT_COUNTS_SUCCESS",
  FETCH_DEPLOYEDPRODUCT_COUNTS_FAILURE = "FETCH_DEPLOYEDPRODUCT_COUNTS_FAILURE",
  CREATE_DEPLOYEDPRODUCT_REQUEST = "CREATE_DEPLOYEDPRODUCT_REQUEST",
  CREATE_DEPLOYEDPRODUCT_SUCCESS = "CREATE_DEPLOYEDPRODUCT_SUCCESS",
  CREATE_DEPLOYEDPRODUCT_FAILURE = "CREATE_DEPLOYEDPRODUCT_FAILURE",
  CREATE_RANGE_DEPLOYEDPRODUCT_REQUEST = "CREATE_RANGE_DEPLOYEDPRODUCT_REQUEST",
  CREATE_RANGE_DEPLOYEDPRODUCT_SUCCESS = "CREATE_RANGE_DEPLOYEDPRODUCT_SUCCESS",
  CREATE_RANGE_DEPLOYEDPRODUCT_FAILURE = "CREATE_RANGE_DEPLOYEDPRODUCT_FAILURE",
  UPDATE_DEPLOYEDPRODUCT_REQUEST = "UPDATE_DEPLOYEDPRODUCT_REQUEST",
  UPDATE_DEPLOYEDPRODUCT_SUCCESS = "UPDATE_DEPLOYEDPRODUCT_SUCCESS",
  UPDATE_DEPLOYEDPRODUCT_FAILURE = "UPDATE_DEPLOYEDPRODUCT_FAILURE",
  REMOVE_DEPLOYEDPRODUCT_REQUEST = "REMOVE_DEPLOYEDPRODUCT_REQUEST",
  REMOVE_DEPLOYEDPRODUCT_SUCCESS = "REMOVE_DEPLOYEDPRODUCT_SUCCESS",
  REMOVE_DEPLOYEDPRODUCT_FAILURE = "REMOVE_DEPLOYEDPRODUCT_FAILURE",
  REMOVE_RANGE_DEPLOYEDPRODUCT_REQUEST = "REMOVE_RANGE_DEPLOYEDPRODUCT_REQUEST",
  REMOVE_RANGE_DEPLOYEDPRODUCT_SUCCESS = "REMOVE_RANGE_DEPLOYEDPRODUCT_SUCCESS",
  REMOVE_RANGE_DEPLOYEDPRODUCT_FAILURE = "REMOVE_RANGE_DEPLOYEDPRODUCT_FAILURE",
  SET_DEPLOYEDPRODUCT = "SET_DEPLOYEDPRODUCT",
  CLEAR_DEPLOYEDPRODUCT = "CLEAR_DEPLOYEDPRODUCT",
  SET_DEPLOYEDPRODUCTS = "SET_DEPLOYEDPRODUCTS",
  CLEAR_DEPLOYEDPRODUCTS = "CLEAR_DEPLOYEDPRODUCTS",
}